refactor code and create utils directory for shared utility functions
