+++ egenerator/manager/reconstruction/modules/reconstruction.py
@@ -1,6 +1,8 @@
 import numpy as np
 import tensorflow as tf
 
+from egenerator.manager.reconstruction.modules.utils import trafo
+
 
 class Reconstruction:
 
@@ -243,12 +245,12 @@ class Reconstruction:
             data_batch, seed_tensor)
 
         # invert possible transformation and put full hypothesis together
-        result = self.get_reco_result_batch(
+        result = trafo.get_reco_result_batch(
             result_trafo=result_trafo,
-            data_batch=data_batch,
             seed_tensor=seed_tensor,
             fit_paramater_list=self.fit_paramater_list,
             minimize_in_trafo_space=self.minimize_in_trafo_space,
+            data_trafo=self.manager.data_trafo,
             parameter_tensor_name=self.parameter_tensor_name)
 
         loss_seed = self.parameter_loss_function(
@@ -265,71 +267,6 @@ class Reconstruction:
             'seed_from_previous_module': self.seed_from_previous_module,
         }
         return result_dict
-
-    def get_reco_result_batch(self, result_trafo,
-                              data_batch,
-                              seed_tensor,
-                              fit_paramater_list,
-                              minimize_in_trafo_space,
-                              parameter_tensor_name='x_parameters'):
-        """Get the reco result batch.
-
-        This inverts a possible transformation if minimize_in_trafo_space is
-        True and also puts the full hypothesis back together if only parts
-        of it were fitted
-
-        Parameters
-        ----------
-        result_trafo : TYPE
-            Description
-        data_batch : TYPE
-            Description
-        seed_tensor : TYPE
-            Description
-        fit_paramater_list : TYPE
-            Description
-        minimize_in_trafo_space : TYPE
-            Description
-        parameter_tensor_name : str, optional
-            Description
-
-        Returns
-        -------
-        tf.Tensor
-            The full result batch.
-
-        """
-        if minimize_in_trafo_space:
-            cascade_seed_batch_trafo = self.manager.data_trafo.transform(
-                        data=seed_tensor,
-                        tensor_name=parameter_tensor_name)
-            try:
-                cascade_seed_batch_trafo = cascade_seed_batch_trafo.numpy()
-            except AttributeError:
-                pass
-        else:
-            cascade_seed_batch_trafo = seed_tensor
-
-        if np.all(fit_paramater_list):
-            cascade_reco_batch = result_trafo
-        else:
-            # get seed parameters
-            cascade_reco_batch = []
-            result_counter = 0
-            for i, fit in enumerate(fit_paramater_list):
-                if fit:
-                    cascade_reco_batch.append(result_trafo[:, result_counter])
-                    result_counter += 1
-                else:
-                    cascade_reco_batch.append(cascade_seed_batch_trafo[:, i])
-            cascade_reco_batch = np.array(cascade_reco_batch).T
-
-        # transform back if minimization was performed in trafo space
-        if minimize_in_trafo_space:
-            cascade_reco_batch = self.manager.data_trafo.inverse_transform(
-                data=cascade_reco_batch,
-                tensor_name=parameter_tensor_name)
-        return cascade_reco_batch
 
 
 class SelectBestReconstruction:
